[perf] probe: Allow weak symbols to be probed
Message-id: <1431774138-14330-173-git-send-email-jolsa@redhat.com>
Patchwork-id: 112393
O-Subject: [PATCH RHEL7.2 BZ1222189 172/327] perf probe: Allow weak symbols to be probed
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit e578da3b2009da2a9ae2d25fd0f78c7b76ca5e56
Author: Namhyung Kim <namhyung@kernel.org>
Date: Fri Mar 6 16:31:29 2015 +0900

description
===========
It currently prevents adding probes in weak symbols.  But there're cases
that given name is an only weak symbol so that we cannot add probe.

  $ perf probe -x /usr/lib/libc.so.6 -a calloc
  Failed to find symbol calloc in /usr/lib/libc-2.21.so
    Error: Failed to add events.

  $ nm /usr/lib/libc.so.6 | grep calloc
  000000000007b1f0 t __calloc
  000000000007b1f0 T __libc_calloc
  000000000007b1f0 W calloc

This change will result in duplicate probes when strong and weak symbols
co-exist in a binary.  But I think it's not a big problem since probes
at the weak symbol will never be hit anyway.

Signed-off-by: Rafael Aquini <aquini@redhat.com>