[perf] tools: Fix perf-read-vdsox32 not building and lib64 install dir
Message-id: <1431774138-14330-202-git-send-email-jolsa@redhat.com>
Patchwork-id: 112422
O-Subject: [PATCH RHEL7.2 BZ1222189 201/327] perf tools: Fix perf-read-vdsox32 not building and lib64 install dir
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 76aea7731e7050c066943a1d7456ec6510702601
Author: H.J. Lu <hjl.tools@gmail.com>
Date: Tue Mar 17 15:27:48 2015 -0700

description
===========
Commit:

  c6e5e9fbc3ea ("perf tools: Fix building error in x86_64 when dwarf unwind is on")

removed the definition of IS_X86_64 but not all places using it, with
the consequence that perf-read-vdsox32 would not be built anymore, and
the default lib install directory was 'lib' instead of 'lib64'.

Also needs to go to v3.19.

Signed-off-by: Rafael Aquini <aquini@redhat.com>