[perf] trace: Handle legacy syscalls tracepoints
Message-id: <1431774138-14330-227-git-send-email-jolsa@redhat.com>
Patchwork-id: 112447
O-Subject: [PATCH RHEL7.2 BZ1222189 226/327] perf trace: Handle legacy syscalls tracepoints
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit f208bd8df08d0659ecd9e09e992c79aab26cddfa
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Mar 20 17:46:53 2015 -0300

description
===========
Currently the code skips the first field with the expectation that it is 'nr'.
But older kernels do not have the 'nr' field:

    field:int nr;   offset:8;   size:4; signed:1;

Change perf-trace to drop the field if it exists after parsing the format file.

This fixes the off-by-one problem with older kernels (e.g., RHEL6). e.g,
perf-trace shows this for write:

  1.515 ( 0.006 ms): dd/4245 write(buf: 2</dev/pts/0>, count: 140733837536224       ) = 26

where 2 is really the fd, the huge number is really the buf address, etc.  With
this patch you get the more appropriate:

  1.813 ( 0.003 ms): dd/6330 write(fd: 2</dev/pts/0>, buf: 0x7fff22fc81f0, count: 25) = 25

Signed-off-by: Rafael Aquini <aquini@redhat.com>