[perf] symbols: Save DSO loading errno to better report errors
Message-id: <1431774138-14330-249-git-send-email-jolsa@redhat.com>
Patchwork-id: 112469
O-Subject: [PATCH RHEL7.2 BZ1222189 248/327] perf symbols: Save DSO loading errno to better report errors
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 18425f13a0890ac1e88a64276771c1ae10030b4a
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Mar 24 11:49:02 2015 -0300

description
===========
Before, when some problem happened while trying to load the kernel
symtab, 'perf top' would show:

      Warning:
      The vmlinux file can't be used.
      Kernel samples will not be resolved.

      Press any key...

Now, it reports:

  # perf top --vmlinux /dev/null

      Warning:
      The /tmp/passwd file can't be used: Invalid ELF file
      Kernel samples will not be resolved.

      Press any key...

This is possible because we now register the reason for not being able
to load the symtab in the dso->load_errno member, and provide a
dso__strerror_load() routine to format this error into a strerror like
string with a short reason for the error while loading.

That can be just forwarding the dso__strerror_load() call to
strerror_r(), or, for a separate errno range providing a custom message.

Signed-off-by: Rafael Aquini <aquini@redhat.com>