[perf] tools lib traceevent: Handle NULL comm name
Message-id: <1431774138-14330-250-git-send-email-jolsa@redhat.com>
Patchwork-id: 112470
O-Subject: [PATCH RHEL7.2 BZ1222189 249/327] tools lib traceevent: Handle NULL comm name
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit deab6f55a2fe6fe044af61c9aad6a8e90cda6499
Author: Josef Bacik <jbacik@fb.com>
Date: Tue Mar 24 09:57:49 2015 -0400

description
===========
It is possible that a pid has no associated comm attached to it, although it
can still be passed to pevent_register_comm().

But if comm is NULL, it will cause strdup() to segfault. To prevent this
from happening, if comm is NULL use the default "<...>" name for the
pid.

Signed-off-by: Rafael Aquini <aquini@redhat.com>