[perf] tools lib traceevent: Copy trace_clock and free it
Message-id: <1431774138-14330-251-git-send-email-jolsa@redhat.com>
Patchwork-id: 112471
O-Subject: [PATCH RHEL7.2 BZ1222189 250/327] tools lib traceevent: Copy trace_clock and free it
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 99ad1417db2e83530b88e3eff1a40285d6ebfee9
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Tue Mar 24 09:57:50 2015 -0400

description
===========
The pevent->trace_clock should not be a direct pointer to what was
given. It should be copied and freed.

Note, valgrind pointed this out when a caller passed in a pointer that
needed to be freed and it never was. Ideally, pevent should copy it
(which this change does), and free the copy. It's up to the caller to
free the clock string passed in.

Signed-off-by: Rafael Aquini <aquini@redhat.com>