[perf] tools lib traceevent: Make plugin options either string or boolean
Message-id: <1431774138-14330-254-git-send-email-jolsa@redhat.com>
Patchwork-id: 112474
O-Subject: [PATCH RHEL7.2 BZ1222189 253/327] tools lib traceevent: Make plugin options either string or boolean
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 5dbcfd930e430b33fb4dfded9c1af0cf432772b9
Author: Steven Rostedt <srostedt@redhat.com>
Date: Tue Mar 24 09:57:54 2015 -0400

description
===========
When a plugin option is defined, by default it is a boolean (true or false).

If the option is something else, then it needs to set its "value" field to
a default string other than NULL (can be just "").

If the value is not set then the option is considered boolean, and the
updating of the option value will be handled accordingly.

Signed-off-by: Rafael Aquini <aquini@redhat.com>