[perf] tools lib traceevent kbuffer: Remove extra update to data pointer in PADDING
Message-id: <1431774138-14330-255-git-send-email-jolsa@redhat.com>
Patchwork-id: 112475
O-Subject: [PATCH RHEL7.2 BZ1222189 254/327] tools lib traceevent kbuffer: Remove extra update to data pointer in PADDING
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit c5e691928bf166ac03430e957038b60adba3cf6c
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Tue Mar 24 09:57:55 2015 -0400

description
===========
When a event PADDING is hit (a deleted event that is still in the ring
buffer), translate_data() sets the length of the padding and also updates
the data pointer which is passed back to the caller.

This is unneeded because the caller also updates the data pointer with
the passed back length. translate_data() should not update the pointer,
only set the length.

Signed-off-by: Rafael Aquini <aquini@redhat.com>