[perf] tools: Add pid/tid filtering to report and script commands
Message-id: <1431774138-14330-260-git-send-email-jolsa@redhat.com>
Patchwork-id: 112480
O-Subject: [PATCH RHEL7.2 BZ1222189 259/327] perf tools: Add pid/tid filtering to report and script commands
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit e03eaa400cf8b8bded86cc5c41018a1c69152f16
Author: David Ahern <dsahern@gmail.com>
Date: Tue Mar 24 09:52:41 2015 -0600

description
===========
The 'record' and 'top' tools already allow a user to specify a CSV of
pids and/or tids of tasks to collect data.

Add those options to the 'report' and 'script' analysis commands to only
consider samples related to the given pids/tids.

This is also inline with the existing comm option.

Signed-off-by: Rafael Aquini <aquini@redhat.com>