[perf] tools lib traceevent: Zero should not be considered "not found" in eval_flag()
Message-id: <1431774138-14330-266-git-send-email-jolsa@redhat.com>
Patchwork-id: 112486
O-Subject: [PATCH RHEL7.2 BZ1222189 265/327] tools lib traceevent: Zero should not be considered "not found" in eval_flag()
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 7c27f78a297b54c3c2f5075cb15d33431b7f6333
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Tue Mar 24 14:58:13 2015 -0400

description
===========
Guilherme Cox found that:

 There is, however, a potential bug if there is an item with code zero
 that is not the first one in the symbol list, since eval_flag(..)
 returns 0 when it doesn't find anything.

That is, if you have the following enums:

enum {
  FOO_START = 0,
  FOO_GO    = 1,
  FOO_END   = 2
}

and then have:

  __print_symbolic(foo, FOO_GO, "go", FOO_START, "start",
		        FOO_END, "end")

If none of the enums are known to pevent, then eval_flag() will return
zero, and it will match it to the first item in the list, which would be
FOO_GO, which is not zero.

Luckily, in most cases, the first element would be zero, and the parsing
would match out of sheer luck.

Signed-off-by: Rafael Aquini <aquini@redhat.com>