[perf] session: Always initialize ordered_events
Message-id: <1431774138-14330-271-git-send-email-jolsa@redhat.com>
Patchwork-id: 112491
O-Subject: [PATCH RHEL7.2 BZ1222189 270/327] perf session: Always initialize ordered_events
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit aae59fab978e143f24d22e52f075c4be3ee4e628
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Mar 31 09:53:50 2015 -0300

description
===========
Even when it is not used to actually reorder events, some of its fields
are used, like session->ordered_events->tool, to shorten function
signatures where tool, for instance, was being passed, as the tool is
needed for the ordered_events code, we need it there and might as well
use it for other perf_session needs.

This fixes a problem where 'perf script' had some condition that made
session->ordered_events not to be initialized even with its
script->tool ordered_events related flags asking for it to be, which
looks like another bug and needs to be investigated further.

Always initializing session->ordered_events at least leaves the current
assumptions in place, so do it now.

Signed-off-by: Rafael Aquini <aquini@redhat.com>