[perf] tools: Fix cross-endian analysis
Message-id: <1431774138-14330-307-git-send-email-jolsa@redhat.com>
Patchwork-id: 112527
O-Subject: [PATCH RHEL7.2 BZ1222189 306/327] perf tools: Fix cross-endian analysis
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 1060ab857f64f33a9445881fd31fa91470aeb622
Author: David Ahern <david.ahern@oracle.com>
Date: Thu Apr 9 16:15:46 2015 -0400

description
===========
Trying to analyze a big endian data file on little endian system fails
with the error:

  0xa9b40 [0x70]: failed to process type: 9

The problem is that header parsing is not done correctly because the
file attributes are not swapped. Make it so. With this patch able to
analyze a sparc64 data file on x86_64.

Signed-off-by: Rafael Aquini <aquini@redhat.com>