[perf] probe: Check the orphaned -x option
Message-id: <1431774138-14330-310-git-send-email-jolsa@redhat.com>
Patchwork-id: 112530
O-Subject: [PATCH RHEL7.2 BZ1222189 309/327] perf probe: Check the orphaned -x option
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit 8cb0aa4c2db395b143cd5165586dc17677960002
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Wed Apr 1 19:25:42 2015 +0900

description
===========
To avoid probing in unintended binary, the orphaned -x option must be
checked and warned.

Without this patch, following command sets up the probe in the kernel.

  -----
  # perf probe -a strcpy -x ./perf
  Added new event:
    probe:strcpy         (on strcpy)

  You can now use it in all perf tools, such as:

          perf record -e probe:strcpy -aR sleep 1
  -----

But in this case, it seems that the user may want to probe in the perf
binary. With this patch, perf-probe correctly handles the orphaned -x.

  -----
  # perf probe -a strcpy -x ./perf
    Error: -x/-m must follow the probe definitions.
  ...
  -----

Signed-off-by: Rafael Aquini <aquini@redhat.com>