[perf] trace: Enable events when doing system wide tracing and starting a workload
Message-id: <1431774138-14330-316-git-send-email-jolsa@redhat.com>
Patchwork-id: 112536
O-Subject: [PATCH RHEL7.2 BZ1222189 315/327] perf trace: Enable events when doing system wide tracing and starting a workload
Bugzilla: 1222189
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1222189
https://bugzilla.redhat.com/show_bug.cgi?id=1222189

upstream
========
commit cb24d01d217497fb32467de22d773655f47d3896
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Apr 22 10:04:23 2015 -0300

description
===========
 commit f7aa222ff397
 Author: Arnaldo Carvalho de Melo <acme@redhat.com>
 Date:   Tue Feb 3 13:25:39 2015 -0300

    perf trace: No need to enable evsels for workload started from perf

The assumption was that whenever a workload is specified, the
attr.enable_on_exec evsel flag would be set, but that is not happening
when perf_record_opts.system_wide is set, for instance

That resulted in both perf_evlist__enable() and attr.enable_on_exec
being not called/set, which made the events to remain disabled while the
workload runs, producing no output.

Fix it,  by calling perf_evlist__enable() in the 'trace' tool
when forking and not targetting a workload started from trace

v2: Test against !target__none(), as suggested by Namhyung Kim, that is
what is used in perf_evsel__config() when deciding if the
attr.enable_on_exec flag to be set. More work is needed to cover other
cases such as opts->initial_delay.

Signed-off-by: Rafael Aquini <aquini@redhat.com>