[fs] xfs: clean up inode locking for RENAME_WHITEOUT
Message-id: <1432045664-23943-10-git-send-email-cmaiolino@redhat.com>
Patchwork-id: 112909
O-Subject: [RHEL7 fs PATCH 09/13] xfs: clean up inode locking for RENAME_WHITEOUT
Bugzilla: 1158888
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Dave Chinner <dchinner@redhat.com>

Upstream: 95afcf5

When doing RENAME_WHITEOUT, we now have to lock 5 inodes into the
rename transaction. This means we need to update
xfs_sort_for_rename() and xfs_lock_inodes() to handle up to 5
inodes. Because of the vagaries of rename, this means we could have
anywhere between 3 and 5 inodes locked into the transaction....

While xfs_lock_inodes() does not need anything other than an assert
telling us we are passing more inodes that we ever thought we should
see, it could do with a logic rework to remove all the indenting.
This is not a functional change - it just makes the code a lot
easier to read.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>
Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>