[block] nvme: Admin queue removal handling
Message-id: <1432147495-21058-45-git-send-email-dmilburn@redhat.com>
Patchwork-id: 113132
O-Subject: [RHEL7.2 PATCH BZ 1179447 44/65] NVMe: Admin queue removal handling
Bugzilla: 1179447
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Keith Busch <keith.busch@intel.com>

This protects admin queue access on shutdown. When the controller is
disabled, the queue is frozen to prevent new entry, and unfrozen on
resume, and fixes cq_vector signedness to not suspend a queue twice.

Since unfreezing the queue makes it available for commands, it requires
the queue be initialized, so this moves this part after that.

Special handling is done when the device is unresponsive during
shutdown. This can be optimized to not require subsequent commands to
timeout, but saving that fix for later.

This patch also removes the kill signals in this path that were left-over
artifacts from the blk-mq conversion and no longer necessary.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 0fb59cbc5f133207535b25ec7d16fba24d549ee2)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>