[fs] xfs: allow linkat() on O_TMPFILE files
Message-id: <1432303680-10221-1-git-send-email-cmaiolino@redhat.com>
Patchwork-id: 113577
O-Subject: [RHEL7 fs PATCH] xfs: allow linkat() on O_TMPFILE files
Bugzilla: 1158888
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Upstream: ab29743

The VFS allows an anonymous temporary file to be named at a later
time via a linkat() syscall. The inodes for O_TMPFILE files are
are marked with a special flag I_LINKABLE and have a zero link count.

To support this in XFS, xfs_link() detects if this flag I_LINKABLE
is set and behaves appropriately when detected. So in this case,
its transaciton reservation takes into account the additional
overhead of removing the inode from the unlinked list. Then the
inode is removed from the unlinked list and the directory entry
is added. Finally its link count is bumped accordingly.

Signed-off-by: Zhi Yong Wu <wuzhy@linux.vnet.ibm.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Ben Myers <bpm@sgi.com>

RHEL7 changes:

XFS_INODE_CLUSTER_SIZE was removed since the upstream patch was written, so, to
make the patch clean, I merged in the patch, the conflicts described in this
upstream merge commit: fe986f9d88ab8079c91669b7f175081f15491a80

Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>