[kvm] x86: Wrong emulation on 'xadd X, X'
Message-id: <1432699356-504-22-git-send-email-bsd@redhat.com>
Patchwork-id: 114234
O-Subject: [PATCH RHEL-7.2 v4 021/312] KVM: x86: Wrong emulation on 'xadd X, X'
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Nadav Amit <namit@cs.technion.ac.il>

The emulator does not emulate the xadd instruction correctly if the two
operands are the same.  In this (unlikely) situation the result should be the
sum of X and X (2X) when it is currently X.  The solution is to first perform
writeback to the source, before writing to the destination.  The only
instruction which should be affected is xadd, as the other instructions that
perform writeback to the source use the extended accumlator (e.g., RAX:RDX).

Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit ee212297cd425620867d4398d55d068c4203768c)
Signed-off-by: Rafael Aquini <aquini@redhat.com>