[kvm] x86: DR6/7.RTM cannot be written
Message-id: <1432699356-504-75-git-send-email-bsd@redhat.com>
Patchwork-id: 114236
O-Subject: [PATCH RHEL-7.2 v4 074/312] KVM: x86: DR6/7.RTM cannot be written
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Nadav Amit <namit@cs.technion.ac.il>

Haswell and newer Intel CPUs have support for RTM, and in that case DR6.RTM is
not fixed to 1 and DR7.RTM is not fixed to zero. That is not the case in the
current KVM implementation. This bug is apparent only if the MOV-DR instruction
is emulated or the host also debugs the guest.

This patch is a partial fix which enables DR6.RTM and DR7.RTM to be cleared and
set respectively. It also sets DR6.RTM upon every debug exception. Obviously,
it is not a complete fix, as debugging of RTM is still unsupported.

Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 6f43ed01e87c8a8dbd8c826eaf0f714c1342c039)
Signed-off-by: Rafael Aquini <aquini@redhat.com>