[kvm] Faults which trigger IO release the mmap_sem
Message-id: <1432699356-504-149-git-send-email-bsd@redhat.com>
Patchwork-id: 114343
O-Subject: [PATCH RHEL-7.2 v4 148/312] kvm: Faults which trigger IO release the mmap_sem
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

When KVM handles a tdp fault it uses FOLL_NOWAIT. If the guest memory
has been swapped out or is behind a filemap, this will trigger async
readahead and return immediately. The rationale is that KVM will kick
back the guest with an "async page fault" and allow for some other
guest process to take over.

If async PFs are enabled the fault is retried asap from an async
workqueue. If not, it's retried immediately in the same code path. In
either case the retry will not relinquish the mmap semaphore and will
block on the IO. This is a bad thing, as other mmap semaphore users
now stall as a function of swap or filemap latency.

This patch ensures both the regular and async PF path re-enter the
fault allowing for the mmap semaphore to be relinquished in the case
of IO wait.

Reviewed-by: Radim Krm <rkrcmar@redhat.com>
Signed-off-by: Andres Lagar-Cavilla <andreslc@google.com>
Acked-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 234b239bea395316d7f78018c672f4a88b3cdf0d)

* RHEL7 Notes *

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	mm/gup.c
	virt/kvm/async_pf.c
changes in mm/gup.c ported to mm/memory.c

Signed-off-by: Bandan Das <bsd@redhat.com>