[kvm] emulator: fix error code for __linearize
Message-id: <1432699356-504-167-git-send-email-bsd@redhat.com>
Patchwork-id: 114230
O-Subject: [PATCH RHEL-7.2 v4 166/312] KVM: emulator: fix error code for __linearize
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

The error code for #GP and #SS is zero when the segment is used to
access an operand or an instruction.  It is only non-zero when
a segment register is being loaded; for limit checks this means
cases such as:

* for #GP, when RIP is beyond the limit on a far call (before the first
instruction is executed).  We do not implement this check, but it
would be in em_jmp_far/em_call_far.

* for #SS, if the new stack overflows during an inter-privilege-level
call to a non-conforming code segment.  We do not implement stack
switching at all.

So use an error code of zero.

Reviewed-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 3606189fa3da6afcad0cbbc9b91e94f1f158da5a)
Signed-off-by: Rafael Aquini <aquini@redhat.com>