[kvm] x86: Wrong flags on CMPS and SCAS emulation
Message-id: <1432699356-504-202-git-send-email-bsd@redhat.com>
Patchwork-id: 114293
O-Subject: [PATCH RHEL-7.2 v4 201/312] KVM: x86: Wrong flags on CMPS and SCAS emulation
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Nadav Amit <namit@cs.technion.ac.il>

CMPS and SCAS instructions are evaluated in the wrong order.  For reference (of
CMPS), see http://www.fermimn.gov.it/linux/quarta/x86/cmps.htm : "Note that the
direction of subtraction for CMPS is [SI] - [DI] or [ESI] - [EDI]. The left
operand (SI or ESI) is the source and the right operand (DI or EDI) is the
destination. This is the reverse of the usual Intel convention in which the
left operand is the destination and the right operand is the source."

Introducing em_cmp_r for this matter that performs comparison in reverse order
using fastop infrastructure to avoid a wrapper function.

Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 5aca3722362689a679419dc9b3e426921c70ef16)
Signed-off-by: Rafael Aquini <aquini@redhat.com>