[kvm] x86: Stack size is overridden by __linearize
Message-id: <1432699356-504-229-git-send-email-bsd@redhat.com>
Patchwork-id: 114465
O-Subject: [PATCH RHEL-7.2 v4 228/312] KVM: x86: Stack size is overridden by __linearize
Bugzilla: 1209995
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Nadav Amit <namit@cs.technion.ac.il>

When performing segmented-read/write in the emulator for stack operations, it
ignores the stack size, and uses the ad_bytes as indication for the pointer
size. As a result, a wrong address may be accessed.

To fix this behavior, we can remove the masking of address in __linearize and
perform it beforehand.  It is already done for the operands (so currently it is
inefficiently done twice). It is missing in two cases:
1. When using rip_relative
2. On fetch_bit_operand that changes the address.

This patch masks the address on these two occassions, and removes the masking
from __linearize.

Note that it does not mask EIP during fetch. In protected/legacy mode code
fetch when RIP >= 2^32 should result in #GP and not wrap-around. Since we make
limit checks within __linearize, this is the expected behavior.

Partial revert of commit 518547b32ab4 (KVM: x86: Emulator does not
calculate address correctly, 2014-09-30).

Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 1c1c35ae4b75d6fc08393a1f73a5a06dc68eebb8)
Signed-off-by: Rafael Aquini <aquini@redhat.com>